Refactor to use Full Stack Components pattern with resource route
